+++ services/httpd/handler.go
@@ -25,6 +25,7 @@ import (
 	"github.com/gogo/protobuf/proto"
 	"github.com/golang/snappy"
 	"github.com/influxdata/flux"
+	"github.com/influxdata/flux/lang"
 	"github.com/influxdata/influxdb"
 	"github.com/influxdata/influxdb/logger"
 	"github.com/influxdata/influxdb/models"
@@ -1148,11 +1149,10 @@ func (h *Handler) servePromRead(w http.ResponseWriter, r *http.Request, user met
 			)
 		}
 	}
-	
 	respond(resp)
 }
 
-func (h *Handler) serveFluxQuery(w http.ResponseWriter, r *http.Request) {
+func (h *Handler) serveFluxQuery(w http.ResponseWriter, r *http.Request, user meta.User) {
 	atomic.AddInt64(&h.stats.FluxQueryRequests, 1)
 	defer func(start time.Time) {
 		atomic.AddInt64(&h.stats.FluxQueryRequestDuration, time.Since(start).Nanoseconds())
@@ -1171,7 +1171,23 @@ func (h *Handler) serveFluxQuery(w http.ResponseWriter, r *http.Request) {
 		}
 	}
 
+	if h.Config.AuthEnabled {
+		ctx = meta.NewContextWithUser(ctx, user)
+	}
+
 	pr := req.ProxyRequest()
+
+	// Logging
+	var (
+		stats flux.Statistics
+		n     int64
+	)
+	if h.Config.FluxLogEnabled {
+		defer func() {
+			h.logFluxQuery(n, stats, pr.Compiler, err)
+		}()
+	}
+
 	q, err := h.Controller.Query(ctx, pr.Compiler)
 	if err != nil {
 		h.httpError(w, err.Error(), http.StatusInternalServerError)
@@ -1201,15 +1217,48 @@ func (h *Handler) serveFluxQuery(w http.ResponseWriter, r *http.Request) {
 		}
 		encoder := pr.Dialect.Encoder()
 		results := flux.NewResultIteratorFromQuery(q)
-		n, err := encoder.Encode(w, results)
+		if h.Config.FluxLogEnabled {
+			if s, ok := results.(flux.Statisticser); ok {
+				defer func() {
+					stats = s.Statistics()
+				}()
+			}
+		}
+		defer results.Release()
+
+		n, err = encoder.Encode(w, results)
 		if err != nil {
-			results.Release()
 			if n == 0 {
 				// If the encoder did not write anything, we can write an error header.
 				h.httpError(w, err.Error(), http.StatusInternalServerError)
 			}
 		}
 	}
+}
+
+func (h *Handler) logFluxQuery(n int64, stats flux.Statistics, compiler flux.Compiler, err error) {
+	var q string
+	switch c := compiler.(type) {
+	case lang.SpecCompiler:
+		q = fmt.Sprint(flux.Formatted(c.Spec))
+	case lang.FluxCompiler:
+		q = c.Query
+	}
+
+	h.Logger.Info("Executed Flux query",
+		zap.String("compiler_type", string(compiler.CompilerType())),
+		zap.Int64("response_size", n),
+		zap.String("query", q),
+		zap.Error(err),
+		zap.Duration("stat_total_duration", stats.TotalDuration),
+		zap.Duration("stat_compile_duration", stats.CompileDuration),
+		zap.Duration("stat_queue_duration", stats.QueueDuration),
+		zap.Duration("stat_plan_duration", stats.PlanDuration),
+		zap.Duration("stat_requeue_duration", stats.RequeueDuration),
+		zap.Duration("stat_execute_duration", stats.ExecuteDuration),
+		zap.Int64("stat_max_allocated", stats.MaxAllocated),
+		zap.Int("stat_concurrency", stats.Concurrency),
+	)
 }
 
 // serveExpvar serves internal metrics in /debug/vars format over HTTP.
@@ -1443,12 +1492,21 @@ type credentials struct {
 	Token    string
 }
 
+func parseToken(token string) (user, pass string, ok bool) {
+	s := strings.IndexByte(token, ':')
+	if s < 0 {
+		return
+	}
+	return token[:s], token[s+1:], true
+}
+
 // parseCredentials parses a request and returns the authentication credentials.
 // The credentials may be present as URL query params, or as a Basic
 // Authentication header.
 // As params: http://127.0.0.1/query?u=username&p=password
 // As basic auth: http://username:password@127.0.0.1
 // As Bearer token in Authorization header: Bearer <JWT_TOKEN_BLOB>
+// As Token in Authorization header: Token <username:password>
 func parseCredentials(r *http.Request) (*credentials, error) {
 	q := r.URL.Query()
 
@@ -1465,11 +1523,22 @@ func parseCredentials(r *http.Request) (*credentials, error) {
 	if s := r.Header.Get("Authorization"); s != "" {
 		// Check for Bearer token.
 		strs := strings.Split(s, " ")
-		if len(strs) == 2 && strs[0] == "Bearer" {
-			return &credentials{
-				Method: BearerAuthentication,
-				Token:  strs[1],
-			}, nil
+		if len(strs) == 2 {
+			switch strs[0] {
+			case "Bearer":
+				return &credentials{
+					Method: BearerAuthentication,
+					Token:  strs[1],
+				}, nil
+			case "Token":
+				if u, p, ok := parseToken(strs[1]); ok {
+					return &credentials{
+						Method:   UserAuthentication,
+						Username: u,
+						Password: p,
+					}, nil
+				}
+			}
 		}
 
 		// Check for basic auth.