+++ tsdb/series_segment.go
@@ -301,12 +301,12 @@ func SplitSeriesOffset(offset int64) (segmentID uint16, pos uint32) {
 	return uint16((offset >> 32) & 0xFFFF), uint32(offset & 0xFFFFFFFF)
 }
 
-// IsValidSeriesSegmentFilename returns true if filename is a 4-character lowercase hexidecimal number.
+// IsValidSeriesSegmentFilename returns true if filename is a 4-character lowercase hexadecimal number.
 func IsValidSeriesSegmentFilename(filename string) bool {
 	return seriesSegmentFilenameRegex.MatchString(filename)
 }
 
-// ParseSeriesSegmentFilename returns the id represented by the hexidecimal filename.
+// ParseSeriesSegmentFilename returns the id represented by the hexadecimal filename.
 func ParseSeriesSegmentFilename(filename string) (uint16, error) {
 	i, err := strconv.ParseUint(filename, 16, 32)
 	return uint16(i), err